+++ src/kernel/cmds/cat.cpp
@@ -1,6 +1,6 @@
 #include "board/mini-uart.hpp"
 #include "cmd.hpp"
-#include "initramfs.hpp"
+#include "fs/initramfs.hpp"
 
 int cmd_cat(int argc, char* argv[]) {
   int r = 0;
