+++ src/kernel/cmds/heap.cpp
@@ -1,6 +1,7 @@
+#include "mm/heap.hpp"
+
 #include "board/mini-uart.hpp"
 #include "cmd.hpp"
-#include "heap.hpp"
 #include "string.hpp"
 
 int cmd_heap(int argc, char* argv[]) {
