--- src/kernel/cpio.cpp
@@ -1,1 +1,0 @@
-#include "cpio.hpp"
