+++ src/kernel/ds/ringbuffer.cpp
@@ -1,4 +1,4 @@
-#include "ringbuffer.hpp"
+#include "ds/ringbuffer.hpp"
 
 #include "util.hpp"
