+++ src/kernel/mm/heap.cpp
@@ -1,4 +1,4 @@
-#include "heap.hpp"
+#include "mm/heap.hpp"
 
 #include "util.hpp"
